Changed type and improved performance
- Type has been changed from Double to Float: useless to use doubles,
especially on Arduino
- changed output calculation in PID_IC in order to improve performance
by reducing number of operations. NOT TESTED, even if should not
present any issue. Old method still present, just commented out

+++ ArduPID/PID.cpp
@@ -16,19 +16,17 @@
 This function should be called inside an If statement that checks if the time 
 elapsed since the last step is greater or equal than the step time
  //---------------------------------------------------------------------------------*/
-void PID::Compute(double e)
-{
-	double Du_p=_kp*e;
-	_Du_i+=_ki*(e+_e_old);
-	_Du_d=_kf*_Du_d+_kd*(e-_e_old);
-    _e_old=e;
-    *_Output=constrain2((Du_p+_Du_i+_Du_d),_outMin,_outMax);
+void PID::Compute(float e) {
+	float Du_p = _kp * e;
+	_Du_i += _ki * (e + _e_old);
+	_Du_d = _kf * _Du_d + _kd * (e - _e_old);
+    _e_old = e;
+    *_Output = constrain2((Du_p + _Du_i + _Du_d), _outMin, _outMax);
 }
 
 /*-------------------------------------Reset---------------------------------------//
 This function can be called everywhere in order to re-initialize the PID
  //---------------------------------------------------------------------------------*/
-void PID::Reset()
-{
+void PID::Reset() {
 PID_def:Reset_def();
 }
